upper case some commands and rename list to brokers
